pyproject.toml
[project]
name = "sefs"
version = "0.1.0"
description = "Semantic Entropy File System — cluster, organize, and chat with your files"
requires-python = ">=3.11,<3.12"
dependencies = [
    # --- Core async + web ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",

    # --- Database ---
    "aiosqlite>=0.20.0",

    # --- Vector store ---
    "faiss-cpu>=1.8.0",

    # --- Numerics (must already be installed before hdbscan/umap) ---
    "numpy>=1.26.0,<2.0.0",
    "scipy>=1.12.0",

    # --- Clustering & projection ---
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.5",
    "scikit-learn>=1.4.0",

    # --- File watching ---
    "watchdog>=4.0.0",

    # --- Utilities ---
    "aiofiles>=23.2.0",
    "chardet>=5.2.0",

    # --- Document/file format support ---
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]
